backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "FastAPI backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "pymongo",
    "python-dotenv",
    "pydantic",
    "pytest",
    "httpx",
    "motor"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
